README.md
It is a simple Number Guessing game in which you have to guess a number between 1 to 10.
First ypu have to deposit Rs.10.
Then you have got 3 chances to guess the correct number. If you guess the correct answer in 1st chance then you will be rewarded with Rs.100. If you guess the correct answer in 2nd chance then you will be rewarded with Rs.50. If you guess the correct answer in 3rd chance then you will be rewarded with Rs.25.
Then it will ask you if you want to play again. If you say YES, then you will be redirected to the game. If you say no, it will ask if you want to withdaraw your money if you have won.
If you want to withdraw your money, say YES and it will ask you the amout to withdraw.
Just enter the amount and the process is finished.